moved modules to lib/
