+++ lib/dspFilter.js
