+++ lib/jensenDiff.js
