+++ lib/smqt.js
